[#1679] Rename session middleware to Onetime::Session
Renames the `Rack::Session::RedisFamilia` middleware to
`Onetime::Session` to better reflect its purpose within the Onetime
application. The underlying functionality and Redis integration remain
the same.
