+++ apps/api/v1/application.rb
@@ -12,8 +12,8 @@ module V1
     @uri_prefix = '/api/v1'
 
     # Session middleware
-    require_relative '../../../lib/rack/session/redis_familia'
-    use Rack::Session::RedisFamilia, {
+    require_relative '../../../lib/onetime/session'
+    use Onetime::Session, {
       expire_after: 86400, # 24 hours
       key: 'ots.session',
       secure: OT.conf&.dig('site', 'ssl') || false,
