+++ apps/web/core/controllers/helpers.rb
@@ -340,7 +340,7 @@ module Core
 
       @request_context_setup = true
 
-      # Session is already loaded by Rack::Session::RedisFamilia middleware
+      # Session is already loaded by Onetime::Session middleware
       # Load customer based on session state
       @cust = current_customer
